Massively added functionality to the schedule page
This includes styling changes, courses can now appear as blocks. A basic functionality where they can (almost) be aligned at the right time, needs fixing however. Courses are added to each day they have as an attribute and are "aligned" to the correct time. Blocks also have color coordination and the height of the blocks are properly represented to the time interval. This also shows up when trying to print. Still needs to be perfect and there are some bugs to fix...